+++ docs/CHANGELOG.md
@@ -119,6 +119,7 @@ Semantic Versioning.
   * Assigning to an enum or namespace no longer crashes quick-lint-js with an
     assertion failure.
   * Subnamespaces can now be named contextual keywords such as `string`.
+  * Import aliases can now be named contextual keywords such as `implements`.
   * Type annotations can now reference types inside namespaces named contextual
     keywords such as `string`.
   * `import A = ns; class A {}` no longer crashes or falsely reports that `A`